Added PID files.
The PID header files are written to be stateless, and work on a pid_data
structure.

Use setPIDConstants to initialize a pid_data structure, then call
updatePID or fixedUpdatePID to get the pid_output.

Use the latter function when using a fixed timestep, and the former when
using a variable timestep.