First implementation of PID.
Added a PWM scaler to relate rad/s to PWM value.

Added other constants for use in the program.

First implementation of the PID loop. It is called in a fixed timestep
interrupt, to make it more efficient. The current velocity is found via
the encoder ticks, and then the error between commanded and current
velocity is computed. This feeds into the PID algorithm.

This is output as the correction in the PWM signal.