Cleaned up code, and corrected PID.
Cleaned up unused defines.

Added some const declarations to make it simpler easier to understand.

Changed the arrow key functionalities to command the velocity ONLY, not
motor direction.

Added directional functionality in the interrupt handler.

Fixed the PID algorithm so it is not just an offset, the PID handles the
whole movement.