Working Serial Communication
Serial communication uses a request-reply loop. The Arduino mega will
request a packet from the line sensor arduino when there is nothing on
the line. Then, it will read it, synchronize using a the
LINE_PACKET_HEADER, then write the data into a struct with the x y and
angular velocities.