Moved to serial debug and reset.
Now, I send a R to reset the line-following mini, and S to start the
line following. At the beginning of every run, we should send an R to
reset and S to start the line following.

'F' now resets the arduino, turns on the led, and then turns off the led
when it is time to start.

'D' just writes serial start and moves to debugging line sensor data.

'R' Sends the reset byte.

'S' Sends the start byte.