Added all the serial terminals to play all games.
Rubiks and Simon are the same controller, and use Serial0.

Serial1 is burned on the system and doesn't work. Debug serial
is moved to this one, and thus we can't simultaneously use the
computer and all systems.

Etch and Card use Serial3.

Etch timeout is 25s.
Rubiks timeout is 10s.
Card timeout is 10s.
Simon timeout is 20s.

Variables for the commands to each of the subsystems have been defined.

Play game commands and timeout commands have been implemented. Early
timeout has been implemented. What we need to fix is ensuring that a late 'D'.