New wiring paradigm.
Photoresistor threshold changed drastically when outdoors, made it
lower.

The motor pins for the front right motor are swapped with the new wiring
design. Changed them in motor_pins.h

The motors on the right side of the robot move in the wrong direction
(they think backwards is forwards). Swapped sign in computeVelocity.